proyecto2

+++ Prot2/src/main/java/modeloMng/TipoExpedienteJpaController.java
@@ -5,17 +5,20 @@
  */
 package modeloMng;
 
-import modeloMng.exceptions.NonexistentEntityException;
 import java.io.Serializable;
+import javax.persistence.Query;
+import javax.persistence.EntityNotFoundException;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Root;
+import modelo.Expediente;
+import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
-import javax.persistence.Query;
-import javax.persistence.EntityNotFoundException;
 import javax.persistence.Persistence;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
 import modelo.TipoExpediente;
+import modeloMng.exceptions.IllegalOrphanException;
+import modeloMng.exceptions.NonexistentEntityException;
 
 /**
  *
@@ -33,11 +36,29 @@ public class TipoExpedienteJpaController implements Serializable {
     }
 
     public void create(TipoExpediente tipoExpediente) {
+        if (tipoExpediente.getExpedienteList() == null) {
+            tipoExpediente.setExpedienteList(new ArrayList<Expediente>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            List<Expediente> attachedExpedienteList = new ArrayList<Expediente>();
+            for (Expediente expedienteListExpedienteToAttach : tipoExpediente.getExpedienteList()) {
+                expedienteListExpedienteToAttach = em.getReference(expedienteListExpedienteToAttach.getClass(), expedienteListExpedienteToAttach.getIdExpediente());
+                attachedExpedienteList.add(expedienteListExpedienteToAttach);
+            }
+            tipoExpediente.setExpedienteList(attachedExpedienteList);
             em.persist(tipoExpediente);
+            for (Expediente expedienteListExpediente : tipoExpediente.getExpedienteList()) {
+                TipoExpediente oldTipoExpedienteOfExpedienteListExpediente = expedienteListExpediente.getTipoExpediente();
+                expedienteListExpediente.setTipoExpediente(tipoExpediente);
+                expedienteListExpediente = em.merge(expedienteListExpediente);
+                if (oldTipoExpedienteOfExpedienteListExpediente != null) {
+                    oldTipoExpedienteOfExpedienteListExpediente.getExpedienteList().remove(expedienteListExpediente);
+                    oldTipoExpedienteOfExpedienteListExpediente = em.merge(oldTipoExpedienteOfExpedienteListExpediente);
+                }
+            }
             em.getTransaction().commit();
         } finally {
             if (em != null) {
@@ -46,12 +67,45 @@ public class TipoExpedienteJpaController implements Serializable {
         }
     }
 
-    public void edit(TipoExpediente tipoExpediente) throws NonexistentEntityException, Exception {
+    public void edit(TipoExpediente tipoExpediente) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            TipoExpediente persistentTipoExpediente = em.find(TipoExpediente.class, tipoExpediente.getIdTipoExpediente());
+            List<Expediente> expedienteListOld = persistentTipoExpediente.getExpedienteList();
+            List<Expediente> expedienteListNew = tipoExpediente.getExpedienteList();
+            List<String> illegalOrphanMessages = null;
+            for (Expediente expedienteListOldExpediente : expedienteListOld) {
+                if (!expedienteListNew.contains(expedienteListOldExpediente)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Expediente " + expedienteListOldExpediente + " since its tipoExpediente field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            List<Expediente> attachedExpedienteListNew = new ArrayList<Expediente>();
+            for (Expediente expedienteListNewExpedienteToAttach : expedienteListNew) {
+                expedienteListNewExpedienteToAttach = em.getReference(expedienteListNewExpedienteToAttach.getClass(), expedienteListNewExpedienteToAttach.getIdExpediente());
+                attachedExpedienteListNew.add(expedienteListNewExpedienteToAttach);
+            }
+            expedienteListNew = attachedExpedienteListNew;
+            tipoExpediente.setExpedienteList(expedienteListNew);
             tipoExpediente = em.merge(tipoExpediente);
+            for (Expediente expedienteListNewExpediente : expedienteListNew) {
+                if (!expedienteListOld.contains(expedienteListNewExpediente)) {
+                    TipoExpediente oldTipoExpedienteOfExpedienteListNewExpediente = expedienteListNewExpediente.getTipoExpediente();
+                    expedienteListNewExpediente.setTipoExpediente(tipoExpediente);
+                    expedienteListNewExpediente = em.merge(expedienteListNewExpediente);
+                    if (oldTipoExpedienteOfExpedienteListNewExpediente != null && !oldTipoExpedienteOfExpedienteListNewExpediente.equals(tipoExpediente)) {
+                        oldTipoExpedienteOfExpedienteListNewExpediente.getExpedienteList().remove(expedienteListNewExpediente);
+                        oldTipoExpedienteOfExpedienteListNewExpediente = em.merge(oldTipoExpedienteOfExpedienteListNewExpediente);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -69,7 +123,7 @@ public class TipoExpedienteJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -80,6 +134,17 @@ public class TipoExpedienteJpaController implements Serializable {
                 tipoExpediente.getIdTipoExpediente();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The tipoExpediente with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            List<Expediente> expedienteListOrphanCheck = tipoExpediente.getExpedienteList();
+            for (Expediente expedienteListOrphanCheckExpediente : expedienteListOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This TipoExpediente (" + tipoExpediente + ") cannot be destroyed since the Expediente " + expedienteListOrphanCheckExpediente + " in its expedienteList field has a non-nullable tipoExpediente field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(tipoExpediente);
             em.getTransaction().commit();
